portal/requirements.txt
flask==3.0.3
flask-cors==4.0.1
werkzeug==3.0.3
python-dotenv==1.0.1
